Refactorización de validadores de formularios: Los validadores se han extraído de la capa de servicios y se han movido a un archivo helper
