+++ src/app/modules/modals/login-modal/login-modal.component.ts
@@ -8,6 +8,7 @@ import { type User } from 'src/app/models/user.model'
 import { ValidatorsService } from 'src/app/services/validators.service'
 import { environment } from 'src/environments/environment'
 import { TranslateService } from '@ngx-translate/core'
+import { isValidDni, isValidInput, getInputError } from '../../../helpers/form-validators.helper'
 
 @Component({
   selector: 'app-login-modal',
@@ -25,7 +26,7 @@ export class LoginModalComponent {
   loginError: string = ''
 
   loginForm = this.formBuilder.group({
-    dni: ['', Validators.required, this.validatorsService.isValidDni],
+    dni: ['', Validators.required, isValidDni],
     password: ['', [Validators.required, Validators.minLength(6)]]
   })
 
@@ -50,7 +51,7 @@ export class LoginModalComponent {
   };
 
   public isValidField (field: string): boolean {
-    return this.validatorsService.isValidInput(field, this.loginForm)
+    return isValidInput(field, this.loginForm)
   };
 
   public openSuccessfulLoginModal (res: any): void {
@@ -84,11 +85,11 @@ export class LoginModalComponent {
   }
 
   isValidInput (input: string): boolean | null {
-    return this.validatorsService.isValidInput(input, this.loginForm)
+    return isValidInput(input, this.loginForm)
   }
 
   getInputError (field: string): string {
-    return this.validatorsService.getInputError(field, this.loginForm)
+    return getInputError(field, this.loginForm, this.translate)
   }
 
   togglePasswordMode (): void {
